refactor: single-port service, remove microservice servers, replace userDirectory with integrations/userService, unify Postman base URLs
