+++ controllers/assignment.controller.js
@@ -1,5 +1,5 @@
 const { BookingAssignment, Booking } = require('../models/bookingModels');
-const { getPassengerById, getDriverById } = require('../services/userDirectory');
+const { getPassengerById, getDriverById } = require('../integrations/userService');
 
 function toBasicUser(u) {
   if (!u) return undefined;
