+++ postman/Booking_Service.postman_environment.json
@@ -2,10 +2,10 @@
   "id": "booking-service-env",
   "name": "Booking Service Local",
   "values": [
-    { "key": "baseUrl", "value": "http://localhost:4001", "enabled": true },
-    { "key": "bookingBase", "value": "http://localhost:4001", "enabled": true },
-    { "key": "discoveryBase", "value": "http://localhost:4002", "enabled": true },
-    { "key": "userproxyBase", "value": "http://localhost:4003", "enabled": true },
+    { "key": "baseUrl", "value": "http://localhost:4000", "enabled": true },
+    { "key": "bookingBase", "value": "http://localhost:4000", "enabled": true },
+    { "key": "discoveryBase", "value": "http://localhost:4000", "enabled": true },
+    { "key": "userproxyBase", "value": "http://localhost:4000", "enabled": true },
     { "key": "token", "value": "", "enabled": true },
     { "key": "adminId", "value": "", "enabled": true },
     { "key": "token", "value": "", "enabled": true },
