+++ postman/Comprehensive_Ride_Service_Environment.postman_environment.json
@@ -4,25 +4,25 @@
   "values": [
     {
       "key": "baseUrl",
-      "value": "http://localhost:4001",
+      "value": "http://localhost:4000",
       "enabled": true,
       "description": "Base URL for the booking service"
     },
     {
       "key": "bookingBase",
-      "value": "http://localhost:4001",
+      "value": "http://localhost:4000",
       "enabled": true,
       "description": "Booking service base URL"
     },
     {
       "key": "discoveryBase",
-      "value": "http://localhost:4002",
+      "value": "http://localhost:4000",
       "enabled": true,
       "description": "Discovery/Pricing service base URL"
     },
     {
       "key": "userproxyBase",
-      "value": "http://localhost:4003",
+      "value": "http://localhost:4000",
       "enabled": true,
       "description": "User proxy service base URL"
     },
